python-dnssec/requirements.txt
certifi==2020.12.5
chardet==4.0.0
dnspython==2.0.0
filelock==3.0.12
idna==2.10
numpy==1.19.4
pandas==1.2.0
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
requests-file==1.5.1
six==1.15.0
tldextract==3.1.0
tqdm==4.55.0
urllib3==1.26.2
